third_party/blink/web_tests/virtual/fenced-frame-mparch/external/wpt/html/cross-origin-embedder-policy/anonymous-iframe/fenced-frame-coep.tentative.https.window-expected.txt

This is a testharness.js-based test.
PASS Check FencedFrame check adherence to COEP within an anonymous frame
FAIL same-origin => blocked. assert_equals: expected "FencedFrame blocked" but got "FencedFrame loaded"
FAIL cross-origin => blocked. assert_equals: expected "FencedFrame blocked" but got "FencedFrame loaded"
FAIL same-origin + coep => blocked. assert_equals: expected "FencedFrame blocked" but got "FencedFrame loaded"
FAIL cross-origin + coep => blocked. assert_equals: expected "FencedFrame blocked" but got "FencedFrame loaded"
PASS same-origin + coep + corp => allowed.
PASS cross-origin + coep + corp => allowed.
Harness: the test ran to completion.
